feat: Redesigned website
fix: Filament Articles

chore: update packages
