+++ composer.json
@@ -11,8 +11,8 @@
         "php": "^8.4",
         "backstage/laravel-og-image": "^1.0.9",
         "based/momentum-trail": "^0.4.0",
-        "filament/filament": "^3.3.26",
-        "filament/spatie-laravel-media-library-plugin": "^3.3.26",
+        "filament/filament": "^3.3.27",
+        "filament/spatie-laravel-media-library-plugin": "^3.3.27",
         "inertiajs/inertia-laravel": "^2.0.3",
         "intervention/image-laravel": "^1.5.6",
         "intervention/validation": "^4.6.0",
