+++ resources/css/app.css
@@ -13,7 +13,11 @@
     body {
         font-family: "Montserrat", sans-serif;
         font-optical-sizing: auto;
-        @apply h-full w-full antialiased dark:bg-base-100 dark:text-base-content;
+        @apply h-full w-full antialiased bg-gray-100 dark:bg-base-100 dark:text-base-content;
+    }
+
+    #app {
+        @apply flex min-h-screen flex-col;
     }
 }
 
@@ -26,7 +30,7 @@
 }
 
 [class*='btn-'] {
-    @apply rounded border border-transparent p-2 px-4 text-center text-sm font-medium focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2;
+    @apply rounded border cursor-pointer border-transparent p-2 px-4 text-center text-sm font-medium focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2;
 }
 
 [class*='btn-']:disabled {
@@ -67,7 +71,7 @@
         height: initial;
         z-index: 1;
 
-        @apply bg-gradient-to-t from-[#fff] dark:from-[#2a303c];
+        @apply bg-linear-to-t from-[var(--color-gray-100)];
     }
 
     & > .hero-copy {
@@ -99,6 +103,14 @@
     }
 }
 
+.dark {
+    .hero-banner {
+        &::before {
+            @apply from-[#2A303C]!;
+        }
+    }
+}
+
 .hexagon {
     clip-path: polygon(25% 0%, 75% 0%, 100% 50%, 75% 100%, 25% 100%, 0% 50%);
     @apply flex items-center justify-center bg-indigo-500 p-2 font-semibold text-white;
@@ -112,6 +124,13 @@
 
 .twitter-tweet.twitter-tweet-rendered {
     @apply mx-auto;
+}
+
+.prose {
+    & > p, & > p > span, & > span {
+        font-family: Roboto, sans-serif !important;
+        @apply text-black dark:text-base-content;
+    }
 }
 
 .prose iframe[allowfullscreen] {
