+++ resources/js/components/LiveStream.vue
@@ -39,9 +39,9 @@ onBeforeUnmount(() => {
 <template>
     <div v-if="isLive" class="mb-4 flex flex-col gap-4 lg:flex-row">
         <div class="flex-1">
-            <div class="aspect-video w-full overflow-hidden rounded">
+            <div class="aspect-video w-full overflow-hidden rounded shadow-sm ring-1 ring-black/5">
                 <iframe
-                    :src="`https://player.twitch.tv/?channel=${activeChannel}&autoplay=true&muted=true&parent=senshudo.tv&parent=staging.senshudo.tv`"
+                    :src="`https://player.twitch.tv/?channel=${activeChannel}&autoplay=true&muted=true&parent=senshudo.tv&parent=staging.senshudo.tv&parent=senshudo.local`"
                     frameborder="0"
                     allowfullscreen="true"
                     scrolling="no"
@@ -51,8 +51,8 @@ onBeforeUnmount(() => {
         </div>
         <iframe
             id="twitch-chat-embed"
-            :src="`https://www.twitch.tv/embed/${activeChannel}/chat?parent=senshudo.tv&parent=staging.senshudo.tv${isDark ? '&darkpopout' : ''}`"
-            class="hidden rounded sm:flex md:min-w-[350px]"
+            :src="`https://www.twitch.tv/embed/${activeChannel}/chat?parent=senshudo.tv&parent=staging.senshudo.tv&parent=senshudo.local${isDark ? '&darkpopout' : ''}`"
+            class="hidden overflow-hidden rounded shadow-sm ring-1 ring-black/5 sm:flex md:min-w-[350px]"
         ></iframe>
     </div>
 </template>
