--- resources/js/pages-old/about.vue
@@ -1,7 +1,0 @@
-<script lang="ts" setup></script>
-
-<template>
-    <AppHead title="About" />
-
-    <div class="prose mx-auto max-w-7xl p-4 pt-8 dark:text-white"></div>
-</template>
