+++ resources/js/pages/embed/age-verification.vue
