+++ resources/js/pages/embed/video.vue
